Refactor exceptions modules
